fix addr ghosting
This fixes the address ghosting issue, where painters were getting the wrong y
address for the first cycle of every row. However, this exposes a bug where
painters are getting y positions for the currently displayed row, when they
should be getting y positions for the next row (as that is the row which will
be displayed on latch + rst). Alternatively, we could fix the design so that we
display the previous row.