fs: remove arbitrary delay
It turns out that the SPRAM requires a couple of cycles to warm up before it's
ready to accept writes. This isn't clearly documented much of everywhere, but
it is available in the general lore if you know to ask the right people. For me
that was tnt and daveshah in the 1bitsquared discord, so thanks to them!